+++ src/client/input/input.go
@@ -17,6 +17,7 @@ const (
 	WindowUnfocus
 	WindowAcquireMouse
 	WindowUnacquireMouse
+	WindowToggleFullscreen
 	WindowQuit
 )